move vite-rsc into lib
