+++ eslint.config.ts
@@ -67,7 +67,7 @@ export default tseslint.config(
   {
     files: [
       'packages/waku/cli.js',
-      'packages/waku/src/vite-rsc/**/*',
+      'packages/waku/src/lib/vite-rsc/**/*',
       'packages/create-waku/cli.js',
       'examples/41_path-alias/**/*.tsx',
     ],
