+++ packages/waku/src/lib/vite-rsc/entry.browser.tsx
