+++ packages/waku/src/lib/vite-rsc/lib/browser-preamble.tsx
@@ -1,0 +1,10 @@
+import { setServerCallback } from '@vitejs/plugin-rsc/browser';
+import { unstable_callServerRsc } from '../../../minimal/client.js';
+setServerCallback(unstable_callServerRsc);
+
+if (import.meta.hot) {
+  import.meta.hot.on('rsc:update', (e) => {
+    console.log('[rsc:update]', e);
+    globalThis.__WAKU_RSC_RELOAD_LISTENERS__?.forEach((l) => l());
+  });
+}
