+++ packages/waku/src/lib/vite-rsc/middleware/handler.ts
@@ -1,0 +1,6 @@
+import type { Middleware } from '../../../config.js';
+import { handleRequest } from '../lib/handler.js';
+
+const handler: Middleware = () => handleRequest;
+
+export default handler;
