--- packages/waku/src/vite-rsc/middleware/handler.ts
@@ -1,6 +1,0 @@
-import type { Middleware } from '../../config.js';
-import { handleRequest } from '../lib/handler.js';
-
-const handler: Middleware = () => handleRequest;
-
-export default handler;
